Adds support for custom commands (#265)
* Adds support for custom commands

Supports up to 10 custom commands, defined in the configuration file.
The current or selected URL can be passed to the command by placing
${url} in the command string.

Default hotkeys are Alt-1 through Alt-0 for triggering the command
with the selected link URL as a parameter, or Alt-Shift-1 through
Alt-Shift-0 for triggering the command with the current page URL
as a parameter. This follows the convention set by the copy page
URL (C) and copy selected URL (c) defaults.

Custom commands must be available in the $PATH or use absolute
paths. Relative paths are not supported, neither are pipes or
shell redirections.

* Fix linting issues

* Fix typo

* Further clarify comment

* Use goroutines for Error/Info calls

* Run go fmt to fix issue from merge

---------

Co-authored-by: makeworld <makeworld@protonmail.com>